Controllers complete, deleted Point3d.cpp, issues #9 #10 closed, added src for CarDuino Nano

+++ src/RoboCPCommon/Point3d.h
@@ -1,7 +1,7 @@
 #pragma once
 
 struct Point3d{
-  float x;
-  float y;
-  float z;
+  short x;
+  short y; 
+  short z;
 };
